+++ src/main/java/roomescape/controller/AuthController.java
@@ -1,4 +1,4 @@
-package roomescape.auth;
+package roomescape.controller;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -9,6 +9,8 @@ import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import roomescape.auth.CookieProvider;
+import roomescape.auth.JwtTokenProvider;
 import roomescape.controller.request.LoginMemberRequest;
 import roomescape.controller.response.CheckLoginUserResponse;
 import roomescape.controller.response.LoginMemberResponse;
