+++ src/main/java/roomescape/controller/MemberViewController.java
@@ -1,0 +1,18 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MemberViewController {
+
+    @GetMapping("/signup")
+    public String signupForm() {
+        return "signup";
+    }
+
+    @GetMapping("/login")
+    public String loginForm() {
+        return "login";
+    }
+}
