+++ src/main/java/roomescape/domain/Member.java
@@ -13,7 +13,7 @@ public class Member {
     private final String password;
 
     public Member(final Long id, final String name, final MemberRole role, final String email, final String password) {
-        if (name.length() < 2 || name.length() > 10) { //TODO
+        if (name.length() < 2 || name.length() > 10) {
             throw new ReservationException("예약자명은 2글자에서 10글자까지만 가능합니다.");
         }
         this.id = id;
